Replace workspace system by library system management. Reduce memory print of all runtime object

+++ FlightReLive/Assets/Scripts/Core/FFmpeg/FFmpegHelper.cs
@@ -90,7 +90,7 @@ namespace FlightReLive.Core.FFmpeg
                             container.Height = height;
                         }
 
-                        if (double.TryParse(match.Groups[3].Value, NumberStyles.Any, CultureInfo.InvariantCulture, out double fps))
+                        if (float.TryParse(match.Groups[3].Value, NumberStyles.Any, CultureInfo.InvariantCulture, out float fps))
                         {
                             container.Frequency = fps;
                         }
